Fixes so it's working again!!
This pushes through a bunch of fixes, and updates the game for the
latest (beta) wyjs output.  The game is now working again!!  A few key
points:

* Needed to include std library as std.js since the build system does
not generate binaries for dependencies at this time.

* Needed to update calls from the view to minesweeper methods because
they now use proper package-based mangling.

* Need to update the call to the type test "is$..." inside view.js as
the mangling scheme has changed.  This is pretty ugly.

However, the bonus is that it does now work.  I think it would be quite
nice to turn this into a proper single-page application, rather than the
strange hybrid it currently it.